Eliminando del todo la versión 2 de ContestList
