+++ frontend/server/nginx.rewrites
@@ -15,7 +15,7 @@ location /preguntas/ {
 rewrite ^/admin/user/([a-zA-Z0-9_+.-]+)/?$ /admin/user.php?username=$1 last;
 rewrite ^/admin/support/?$ /admin/support.php last;
 rewrite ^/arena/admin/?$ /arena/admin.php last;
-rewrite ^/arena/?$ /arena/indexv2.php?page_size=10 last;
+rewrite ^/arena/?$ /arena/index.php?page_size=10 last;
 rewrite ^/arena/problem/([a-zA-Z0-9_+-]+)/?$ /arena/problem.php?problem_alias=$1 last;
 rewrite ^/arena/problem/([a-zA-Z0-9_+-]+)/print/?$ /arena/problemprint.php?problem_alias=$1 last;
 rewrite ^/arena/([a-zA-Z0-9_+-]+)/virtual/?$ /arena/virtual.php?contest_alias=$1  last;
