+++ frontend/www/js/omegaup/api_types.ts
@@ -734,45 +734,6 @@ export namespace types {
     ): types.ContestListPayload {
       return ((x) => {
         x.contests = ((x) => {
-          if (x instanceof Object) {
-            Object.keys(x).forEach(
-              (y) =>
-                (x[y] = ((x) => {
-                  if (!Array.isArray(x)) {
-                    return x;
-                  }
-                  return x.map((x) => {
-                    x.finish_time = ((x: number) => new Date(x * 1000))(
-                      x.finish_time,
-                    );
-                    x.last_updated = ((x: number) => new Date(x * 1000))(
-                      x.last_updated,
-                    );
-                    x.original_finish_time = ((x: number) =>
-                      new Date(x * 1000))(x.original_finish_time);
-                    x.start_time = ((x: number) => new Date(x * 1000))(
-                      x.start_time,
-                    );
-                    return x;
-                  });
-                })(x[y])),
-            );
-          }
-          return x;
-        })(x.contests);
-        return x;
-      })(
-        JSON.parse(
-          (document.getElementById(elementId) as HTMLElement).innerText,
-        ),
-      );
-    }
-
-    export function ContestListv2Payload(
-      elementId: string = 'payload',
-    ): types.ContestListv2Payload {
-      return ((x) => {
-        x.contests = ((x) => {
           x.current = ((x) => {
             if (!Array.isArray(x)) {
               return x;
@@ -3092,13 +3053,6 @@ export namespace types {
   }
 
   export interface ContestListPayload {
-    contests: types.TimeTypeContests;
-    countContests: { [key: string]: number };
-    isLogged: boolean;
-    query?: string;
-  }
-
-  export interface ContestListv2Payload {
     contests: types.ContestList;
     countContests: { current: number; future: number; past: number };
     query?: string;
